feat(DR-79): add generic VR environment routing for Barcelona
Extract ParisEnvironment logic into reusable VRDestinationEnvironment
component that accepts any environment config as a prop. Update App.js
routing to use getVREnvironment() registry lookup, supporting Paris,
Barcelona, and any future destination automatically via URL param.
