+++ panorama/src/components/ParisEnvironment.js
@@ -3,141 +3,21 @@
  *
  * Ticket: DR-74 (VR-003 - Environnement VR Paris)
  *
- * Orchestration de toutes les scènes VR de Paris avec navigation
+ * Wrapper rétro-compatible autour de VRDestinationEnvironment
+ * pour l'environnement Paris.
  */
 
-import React, { useState, useCallback, useEffect } from 'react';
+import React from 'react';
 import { Text } from '@react-three/drei';
-import VRScene from './VRScene';
-import useHotspotPanel from '../hooks/useHotspotPanel';
 import parisEnvironment from '../data/paris-environment';
+import VRDestinationEnvironment from './VRDestinationEnvironment';
 
 /**
  * Composant ParisEnvironment
- * Gère la navigation entre les différentes scènes de Paris
+ * Wrapper rétro-compatible qui délègue à VRDestinationEnvironment
  */
 function ParisEnvironment() {
-  const [currentSceneId, setCurrentSceneId] = useState(parisEnvironment.defaultScene);
-  const [currentScene, setCurrentScene] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [sceneHistory, setSceneHistory] = useState([]);
-  const [hotspotInfo, setHotspotInfo] = useState(null);
-
-  // Charger la scène actuelle depuis les données
-  useEffect(() => {
-    const scene = parisEnvironment.scenes.find(s => s.id === currentSceneId);
-
-    if (!scene) {
-      console.error(`❌ Scène introuvable: ${currentSceneId}`);
-      return;
-    }
-
-    console.log(`📍 Changement de scène: ${scene.name}`);
-    setCurrentScene(scene);
-    setLoading(false);
-
-  }, [currentSceneId]);
-
-  // Changer de scène (navigation)
-  const handleSceneChange = useCallback((targetSceneId) => {
-    console.log(`🚀 Navigation: ${currentSceneId} → ${targetSceneId}`);
-
-    // Ajouter la scène actuelle à l'historique
-    setSceneHistory(prev => [...prev, currentSceneId]);
-
-    // Effet de transition (fade)
-    setLoading(true);
-
-    // Changer de scène après un court délai pour la transition
-    setTimeout(() => {
-      setCurrentSceneId(targetSceneId);
-      setHotspotInfo(null); // Réinitialiser l'info hotspot
-    }, 500);
-
-  }, [currentSceneId]);
-
-  // Revenir à la scène précédente
-  const handleGoBack = useCallback(() => {
-    if (sceneHistory.length === 0) {
-      console.log('📍 Déjà à la première scène');
-      return;
-    }
-
-    const previousSceneId = sceneHistory[sceneHistory.length - 1];
-    console.log(`⬅️ Retour à: ${previousSceneId}`);
-
-    setSceneHistory(prev => prev.slice(0, -1));
-    setCurrentSceneId(previousSceneId);
-
-  }, [sceneHistory]);
-
-  // Interaction avec un hotspot
-  const handleHotspotClick = useCallback((hotspot) => {
-    console.log(`📌 Interaction hotspot: ${hotspot.title} (${hotspot.type})`);
-
-    if (hotspot.type === 'info') {
-      setHotspotInfo(hotspot);
-
-      // Masquer l'info après 10 secondes
-      setTimeout(() => {
-        setHotspotInfo(null);
-      }, 10000);
-    } else if (hotspot.type === 'teleport') {
-      // Téléportation vers une autre scène
-      console.log(`🚀 Téléportation vers: ${hotspot.targetScene}`);
-      handleSceneChange(hotspot.targetScene);
-    }
-
-  }, [handleSceneChange]);
-
-  // Utiliser le hook pour gérer le panneau de hotspots
-  useHotspotPanel(currentScene?.hotspots, handleHotspotClick);
-
-  if (!currentScene) {
-    return (
-      <Text
-        position={[0, 1.6, -3]}
-        fontSize={0.3}
-        color="#FF0000"
-        anchorX="center"
-        anchorY="middle"
-      >
-        ❌ Erreur: Scène introuvable
-      </Text>
-    );
-  }
-
-  return (
-    <>
-      {/* Scène VR actuelle */}
-      <VRScene
-        scene={currentScene}
-        onSceneChange={handleSceneChange}
-        onHotspotClick={handleHotspotClick}
-      />
-
-      {/* UI de navigation */}
-      <NavigationUI
-        currentScene={currentScene}
-        canGoBack={sceneHistory.length > 0}
-        onGoBack={handleGoBack}
-        environment={parisEnvironment}
-      />
-
-      {/* Affichage des informations hotspot */}
-      {hotspotInfo && (
-        <HotspotInfoPanel
-          hotspot={hotspotInfo}
-          onClose={() => setHotspotInfo(null)}
-        />
-      )}
-
-      {/* Indicateur de chargement pendant transition */}
-      {loading && (
-        <TransitionOverlay />
-      )}
-    </>
-  );
+  return <VRDestinationEnvironment environment={parisEnvironment} />;
 }
 
 /**
